+++ videoanalyst/pipeline/utils/__init__.py
@@ -2,5 +2,5 @@
 
 from .bbox import (clip_bbox, cxywh2xywh, cxywh2xyxy, xywh2cxywh, xywh2xyxy,
                    xyxy2cxywh, xyxy2xywh)
-from .crop import get_axis_aligned_bbox, get_crop, get_subwindow_tracking
+from .crop import get_axis_aligned_bbox, get_crop, get_crop_pp, get_subwindow_tracking
 from .misc import imarray_to_tensor, tensor_to_numpy